requirements.txt
beautifulsoup4==4.5.1
certifi==2016.9.26
future==0.16.0
Pillow==3.4.2
python-telegram-bot==5.2.0
redis==2.10.5
requests==2.12.1
urllib3==1.19.1
